+++ src/gwt/src/org/rstudio/studio/client/RStudioGinjector.java
@@ -114,7 +114,7 @@ import org.rstudio.studio.client.workbench.prefs.views.PythonPreferencesPaneBase
 import org.rstudio.studio.client.workbench.snippets.SnippetHelper;
 import org.rstudio.studio.client.workbench.snippets.ui.EditSnippetsDialog;
 import org.rstudio.studio.client.workbench.ui.ConsoleTabPanel;
-import org.rstudio.studio.client.workbench.ui.polyfill.FocusVisiblePolyfill;
+import org.rstudio.studio.client.workbench.ui.polyfill.FocusVisibleStyles;
 import org.rstudio.studio.client.workbench.views.connections.ui.ConnectionCodePanel;
 import org.rstudio.studio.client.workbench.views.connections.ui.ConnectionExplorer;
 import org.rstudio.studio.client.workbench.views.connections.ui.NewConnectionInstallOdbcHost;
@@ -400,7 +400,7 @@ public interface RStudioGinjector extends Ginjector
    SessionOpener getSessionOpener();
    VirtualConsoleFactory getVirtualConsoleFactory();
    JobItemFactory getJobItemFactory();
-   FocusVisiblePolyfill getFocusVisiblePolyfill();
+   FocusVisibleStyles getFocusVisibleStyles();
    AriaLiveService getAriaLiveService();
 
    // Pro-only below here