+++ src/gwt/src/org/rstudio/studio/client/common/satellite/Satellite.java
@@ -69,8 +69,8 @@ public class Satellite implements HasCloseHandlers<Satellite>
       // load MathJax
       MathJaxLoader.ensureMathJaxLoaded();
 
-      // load focus-visible polyfill
-      RStudioGinjector.INSTANCE.getFocusVisiblePolyfill().load(null);
+      // load focus-visible css
+      RStudioGinjector.INSTANCE.getFocusVisibleStyles().load(null);
 
       // NOTE: Desktop doesn't seem to get onWindowClosing events in Qt 4.8
       // so we instead rely on an explicit callback from the desktop frame